+++ settings.py
@@ -61,17 +61,22 @@ def get_presets_enum(export_format):
     return presets
 
 
-mode_bundle_types = [('NAME', 'Name', "Bundle by matching object names", 'SYNTAX_OFF', 0),
-                     ('PARENT', 'Parent', "Bundle by the parent object", 'FILE_PARENT', 1),
-                     ('COLLECTION', 'Collection', "Bundle by 'Collections'", 'GROUP', 2),
-                     ('SCENE', 'Scene', "Bundle by current scene", 'SCENE_DATA', 3)]
+mode_bundle_types = [
+    ('NAME', 'Name', "Bundle by matching object names", 'SYNTAX_OFF', 0),
+    ('PARENT', 'Parent', "Bundle by the parent object", 'FILE_PARENT', 1),
+    ('COLLECTION', 'Collection', "Bundle by 'Collections'", 'GROUP', 2),
+    ('SCENE', 'Scene', "Bundle by current scene", 'SCENE_DATA', 3)
+]
 
-mode_pivot_types = [('OBJECT_FIRST', 'First Name', "Pivot at the first object sorted by name", 'FILE_TEXT', 1),
-                    ('OBJECT_LOWEST', 'Lowest Object', "Pivot at the lowest Z object's pivot", 'SORT_ASC', 2),
-                    ('BOUNDS_BOTTOM', 'Bottom Center', "Pivot at the bottom center of the bounds of the bundle", 'AXIS_TOP', 3),
-                    ('SCENE', 'Scene 0,0,0', "Pivot at the Scene center 0,0,0'", 'GRID', 4),
-                    ('PARENT', 'Parent', "Pivot from the parent object", 'FILE_PARENT', 5),
-                    ('EMPTY', 'Empty Gizmo', "Empty gizmo object of: Arrow, Plain Axes, Single Arrow>; global for all bundles (must be selected)", 'EMPTY_AXIS', 6),
-                    ('EMPTY_LOCAL', 'Empty Local Gizmo', "You need to have an empty of type Arrow, Plain Axes or Single Arrow located inside the bundle and its name needs to start with 'pivot'; for example 'pivot.001'", 'EMPTY_ARROWS', 7)]
+mode_pivot_types = [
+    ('OBJECT_FIRST', 'First Name', "Pivot at the first object sorted by name", 'FILE_TEXT', 1),
+    ('OBJECT_LOWEST', 'Lowest Object', "Pivot at the lowest Z object's pivot", 'SORT_ASC', 2),
+    ('BOUNDS_BOTTOM', 'Bottom Center', "Pivot at the bottom center of the bounds of the bundle", 'AXIS_TOP', 3),
+    ('SCENE', 'Scene 0,0,0', "Pivot at the Scene center 0,0,0'", 'GRID', 4),
+    ('PARENT', 'Parent', "Pivot from the parent object", 'FILE_PARENT', 5),
+    ('EMPTY', 'Empty Gizmo', "Empty gizmo object of: Arrow, Plain Axes, Single Arrow>; global for all bundles (must be selected)", 'EMPTY_AXIS', 6),
+    ('EMPTY_LOCAL', 'Empty Local Gizmo', "You need to have an empty of type Arrow, Plain Axes or Single Arrow located inside the bundle and its name needs to start with 'pivot'; for example 'pivot.001'", 'EMPTY_ARROWS', 7),
+    ('COLLECTION', 'Collection', "Pivot from the collection instance offset parameter (Object Properties->Collection->X,Y,Z)", 'GROUP', 8)
+]
 
 debug = False